+++ src/Ext/SWType/Body.cpp
@@ -55,6 +55,7 @@ void SWTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->LimboDelivery_RollChances)
 		.Process(this->LimboKill_AffectsHouse)
 		.Process(this->LimboKill_IDs)
+		.Process(this->LimboKill_Counts)
 		.Process(this->RandomBuffer)
 		.Process(this->Detonate_Warhead)
 		.Process(this->Detonate_Weapon)
@@ -68,6 +69,7 @@ void SWTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->SW_Next_RandomWeightsData)
 		.Process(this->SW_Next_RollChances)
 		.Process(this->ShowTimer_Priority)
+		.Process(this->ShowTimer_Percentage)
 		.Process(this->Convert_Pairs)
 		.Process(this->ShowDesignatorRange)
 		.Process(this->TabIndex)
@@ -143,6 +145,7 @@ void SWTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 	this->LimboDelivery_RollChances.Read(exINI, pSection, "LimboDelivery.RollChances");
 	this->LimboKill_AffectsHouse.Read(exINI, pSection, "LimboKill.AffectsHouse");
 	this->LimboKill_IDs.Read(exINI, pSection, "LimboKill.IDs");
+	this->LimboKill_Counts.Read(exINI, pSection, "LimboKill.Counts");
 	this->SW_Next.Read(exINI, pSection, "SW.Next");
 	this->SW_Next_RealLaunch.Read(exINI, pSection, "SW.Next.RealLaunch");
 	this->SW_Next_IgnoreInhibitors.Read(exINI, pSection, "SW.Next.IgnoreInhibitors");
@@ -150,6 +153,7 @@ void SWTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 	this->SW_Next_RollChances.Read(exINI, pSection, "SW.Next.RollChances");
 
 	this->ShowTimer_Priority.Read(exINI, pSection, "ShowTimer.Priority");
+	this->ShowTimer_Percentage.Read(exINI, pSection, "ShowTimer.Percentage");
 
 	this->EMPulse_WeaponIndex.Read(exINI, pSection, "EMPulse.WeaponIndex");
 	this->EMPulse_SuspendOthers.Read(exINI, pSection, "EMPulse.SuspendOthers");