+++ src/Ext/TerrainType/Hooks.cpp
@@ -39,6 +39,14 @@ DEFINE_HOOK(0x71C84D, TerrainClass_AI_Animated, 0x6)
 				for (int i = 0; i < cellCount; i++)
 					pCell->SpreadTiberium(true);
 
+				const int particleIdx = pTypeExt->SpawnsTiberium_Particle;
+
+				if (particleIdx >= 0)
+				{
+					const auto particleSys = Make_Global<ParticleSystemClass*>(0xA8ED78);
+					particleSys->SpawnParticle(ParticleTypeClass::Array[particleIdx], pThis->Location);
+				}
+
 				// Unset context for CellClass hooks.
 				TerrainTypeTemp::pCurrentType = nullptr;
 				TerrainTypeTemp::pCurrentExt = nullptr;